+++ packages/gallery/src/common/constants/index.js
@@ -25,6 +25,8 @@ import dimensions from './dimensions';
 import gallerySizeType from './gallerySizeType';
 import isVertical from './isVertical';
 import galleryTextAlign from './galleryTextAlign.js';
+import gridStyle from './gridStyle';
+import layoutDirection from './layoutDirection';
 
 export default {
   events,
@@ -53,5 +55,7 @@ export default {
   dimensions,
   gallerySizeType,
   isVertical,
-  galleryTextAlign
+  galleryTextAlign,
+  gridStyle,
+  layoutDirection,
 };