+++ packages/gallery/src/components/gallery/presets/collageGallery.js
@@ -11,13 +11,12 @@ export const fixedStyles = {
   cubeImages: false,
   titlePlacement: PLACEMENTS.SHOW_ON_HOVER,
   groupSize: 3,
-  
+  hasThumbnails: false,
+  groupTypes: '1,2h,2v,3t,3b,3l,3r',
   //this params were moved from the presets in layoutHelper and were not tested and checked yet.
   showArrows: false,
-  groupTypes: '1,2h,2v,3t,3b,3l,3r',
   gallerySize: 0,
   fixedColumns: 0,
-  hasThumbnails: false,
   enableScroll: true,
   isGrid: false,
   isSlider: false,