+++ packages/gallery/src/settings/index.js
@@ -33,6 +33,16 @@ import textBoxWidth from './options/textBoxWidth';
 import imageInfoType from './options/imageInfoType';
 import groupSize from './options/groupSize';
 import collageDensity from './options/collageDensity';
+import gridStyle from './options/gridStyle'
+import hasThumbnails from './options/hasThumbnails';
+import groupTypes from './options/groupTypes';
+import thumbnailSize from './options/thumbnailSize';
+import galleryThumbnailsAlignment from './options/galleryThumbnailsAlignment';
+import isRTL from './options/isRTL';
+import scrollSnap from './options/scrollSnap';
+import itemBorderWidth from './options/itemBorderWidth';
+import itemBorderRadius from './options/itemBorderRadius';
+import itemBorderColor from './options/itemBorderColor';
 import arrowsSize from './options/arrowsSize';
 
 export default {
@@ -71,5 +81,15 @@ export default {
     imageInfoType,
     groupSize,
     collageDensity,
+    gridStyle,
+    hasThumbnails,
+    groupTypes,
+    thumbnailSize,
+    galleryThumbnailsAlignment,
+    isRTL,
+    scrollSnap,
+    itemBorderWidth,
+    itemBorderRadius,
+    itemBorderColor,
     arrowsSize,
 }