+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [v1.7.25](https://pro-gallery-1-7-25.surge.sh) (04/03/2020)
+ 
+#### GALLERY
+ - Add clickTarget to EVENTS.ITEM_CLICKED eventData (options: 'item-media', 'item-info', 'item-container')
+
+---
+## [v1.7.24](https://pro-gallery-1-7-24.surge.sh) (03/03/2020)
+ 
+#### GALLERY
+ -  Fixed classNames in horizontal galleries.
+ -  Changed console.warn to console.log when local
+
+---
 ## [v1.7.23](https://pro-gallery-1-7-23.surge.sh) (25/02/2020)
  
 #### GALLERY